1. Changed Java class activities names to contain the word 'Activity'. Also changed 'OfferItemActivity' to 'GiveItemActivity' 2. Added the 'thank you' popup dialog after pressing 'Give'. Also created a product. 3. Added missing implementation in gradle.build
