+++ app/build.gradle
@@ -28,6 +28,7 @@ dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.firebase:firebase-storage:19.1.0'
     implementation 'com.google.android.gms:play-services-maps:17.0.0'
     implementation 'com.google.android.gms:play-services-location:17.0.0'
     implementation 'androidx.preference:preference:1.1.0-alpha05'
@@ -40,6 +41,8 @@ dependencies {
     implementation 'com.google.firebase:firebase-firestore:21.3.1'
     implementation 'com.firebaseui:firebase-ui-firestore:5.0.0'
     implementation 'com.firebaseui:firebase-ui-auth:4.3.2'
+    implementation 'com.firebase:firebase-client-android:2.5.0'
+    implementation 'com.facebook.android:facebook-android-sdk:5.0.0'
     implementation 'com.twitter.sdk.android:twitter-core:3.1.1'
     implementation 'androidx.multidex:multidex:2.0.1'
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
